Backend/firebase/firebase_init.py
import firebase_admin
from firebase_admin import credentials

try:
    firebase_admin.get_app()
except ValueError:
    cred = credentials.Certificate("firebase-key.json")
    firebase_admin.initialize_app(cred)
